mechanism: assembled-minecart terrain collision + wider suction pull-range
Terrain collision for assembled mechanism minecarts (Part B, B1+B2). An
assembled cart's maxSpeed is now a per-tick clamp to the measured gap between
the structure's leading faces and the nearest wall, instead of a flat 0.4 — NMS
clamps a cart's per-tick displacement to maxSpeed, so the brake is the pin: the
structure decelerates to a flush stop and never clips terrain.

- New CartCollision helper. Gap scan is per collision block (min over blocks),
  anchored on the cart's live location, gated on hasCollision so no-collider
  decorative blocks pass through and sub-cube colliders are scanned full-cell.
  Flush-stop at gap <= 0.04; at-rest cap only when a wall is near (open
  powered-rail launches aren't throttled); curve = both-axis real-cell scan;
  ascending rail also scans the forward+ceiling cells; unloaded chunk = wall.
- Fragile blocks in the path are broken only through a BlockBreakEvent
  attributed to a rider, so claim protection can veto; no rider -> the cart
  stops at the vegetation instead of griefing. The break is a separate
  once-per-tick step, kept out of the pure gap measurement.
- B2 guarded re-snap backstop teleports the cart out of any penetration
  (unsigned-clamp ratchet, curve overshoot, griefed block-place) along the rail
  tangent on slopes; never fires in routine operation.
- Resume kick pushes a docked cart back into motion once its path reopens, and
  disarms once it departs under its own power.
- MechanismMinecartManager wires this into tick() for assembled carts; the
  collision clamp is now the sole maxSpeed writer for them (minecarts-v2.md
  amended so the future train system must route maxSpeed through it).

Cross-mechanism collision (cart<->cart, piston<->mech) is deferred to its own
round (B3).

Suction hopper: pull-range is now a double (default 2.5, pull-box side
2*(range+0.75) -> 6.5^3) instead of an int Chebyshev radius of 1.

+++ src/main/java/anon/def9a2a4/corelib/RotationBlocks.java
@@ -1004,9 +1004,10 @@ final class RotationBlocks {
      * ({@code MechanismRotationDriver}, which passes the live cell and the travelling inventory).
      */
     static void suctionEffect(World world, Location center, int cellX, int cellY, int cellZ,
-                              Inventory internal, int pullRange, double pullStrength) {
-        // Half-extent = full cells (pullRange + 0.5) + a 0.25 margin per side, so pullRange 1 →
-        // 1.75 → 3.5×3.5×3.5 box (the 3×3×3 cells plus a quarter-block reach on every face).
+                              Inventory internal, double pullRange, double pullStrength) {
+        // Half-extent = pullRange + a 0.25 margin per side beyond the cell reach, so the pull box
+        // side is 2*(pullRange + 0.75): pullRange 1 → 3.5³, 2.5 → 6.5³. Item capture still only
+        // happens in the hopper's own 1×1×1 cell below; this box governs the inward pull.
         double r = pullRange + 0.75;
         var box = org.bukkit.util.BoundingBox.of(center.toVector(), r, r, r);
         for (org.bukkit.entity.Entity e : world.getNearbyEntities(box)) {
@@ -1167,7 +1168,7 @@ final class RotationBlocks {
     private static double fanMaxPush;
     private static double fanPushPerPower;
     private static int suctionTickInterval;
-    private static int suctionPullRange;
+    private static double suctionPullRange;
     private static double suctionPullStrength;
 
     /** Staged-break progress against one target material. Package-visible: the mechanism rotation
